Added AbstractEntity
