+++ server/helping-hands-backend/src/main/java/com/helpinghands/backendPrototype/Models/AbstractEntity.java
@@ -1,0 +1,37 @@
+package com.helpinghands.backendPrototype.Models;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
+
+@MappedSuperclass
+public class AbstractEntity {
+    public @GeneratedValue @Id Long id;
+    public String name;
+    public String description;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+}
