+++ server/helping-hands-backend/src/main/java/com/helpinghands/backendPrototype/Models/User.java
@@ -6,33 +6,33 @@ import javax.persistence.Id;
 import java.util.Objects;
 
 @Entity
-public class User {
-    private @Id @GeneratedValue Long id;
-    private String name;
+public class User extends AbstractEntity {
+//    private @Id @GeneratedValue Long id;
+//    private String name;
     private String location; //make a location Enum class?
 
     public User() {}
 
-    public User(String name, String location) {
-        this.name=name;
+    public User(String location) {
+//        this.name=name;
         this.location = location;
     }
 
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
+//    public Long getId() {
+//        return id;
+//    }
+//
+//    public void setId(Long id) {
+//        this.id = id;
+//    }
+//
+//    public String getName() {
+//        return name;
+//    }
+//
+//    public void setName(String name) {
+//        this.name = name;
+//    }
 
     public String getLocation() {
         return location;
